Kajima_Fisheye/requirements.txt
bson==0.5.10
Cython
easydict
matplotlib
mysql-connector==2.2.9
mysql-connector-python==8.0.30
mxnet-cu114
numpy
opencv-python
paho-mqtt==1.6.1
pandas
pyroomacoustics
rapid
redis==4.1.4
scikit-image
scikit-learn
scipy
termcolor
torchvision
webrtcvad